birthdayをint型に変更

+++ app/src/main/java/org/t_robop/y_ogawara/tancolle/Data.java
@@ -16,7 +16,7 @@ public class Data {
     //ふりがな
     private String kana;
     //誕生日
-    private long birthday;
+    private int birthday;
     //カテゴリ
     private String category;
     //twitterのID
@@ -78,7 +78,7 @@ public class Data {
         return birthday;
     }
 
-    public void setBirthday(long birthday) {
+    public void setBirthday(int birthday) {
         this.birthday = birthday;
     }
